Rename glob pattern to wildcard
